extract main.dart widgets
