+++ packages/stream_chat_v1/lib/group_info_screen.dart
@@ -10,8 +10,8 @@ import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 import 'channel_file_display_screen.dart';
 import 'channel_media_display_screen.dart';
+import 'channel_page.dart';
 import 'chat_info_screen.dart';
-import 'main.dart';
 import 'pinned_messages_screen.dart';
 import 'routes/routes.dart';
