+++ packages/stream_chat_v1/lib/new_chat_screen.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import 'channel_page.dart';
 import 'chips_input_text_field.dart';
-import 'main.dart';
 import 'routes/routes.dart';
 
 class NewChatScreen extends StatefulWidget {
